push code

+++ examples/tf-lite/src/main/java/com/anymore/tensorflow/image/ImageCompressor.kt
@@ -37,7 +37,7 @@ internal object ImageCompressor {
     }
 
     fun process(bitmap: Bitmap,inputSize: Size): TensorImage {
-        val tfImage = TensorImage(DataType.UINT8).apply { load(bitmap) }
+        val tfImage = TensorImage(DataType.FLOAT32).apply { load(bitmap) }
         Timber.d("src:w=${tfImage.width},h=${tfImage.height},target:w=${inputSize.width},h=${inputSize.height}")
         val processor = createOrGetProcessor(Size(bitmap.width,bitmap.height),inputSize)
         return processor.process(tfImage)
